Work around bgm playback issues in AkibaUU (333980)
The game has a weird implementation of IMFBytestream, where only the
async BeginRead/EndRead or the sync Read functions work properly,
depending whether or not it has opened the media file in overlapped
mode. If Read is used when the file is opened in overlapped mode,
nothing is read from the file and MF ends up reading uninitialized junk.

The game happens to set a bogus 0xFFFFFFFF pcbRead when the sync
Read fails, which we use to detect if we should try the async api
instead.